+++ src/flow_solver/flow_solver.h
@@ -20,7 +20,8 @@
 #include "physics/physics_factory.h"
 #include "dg/dg.h"
 #include "dg/dg_factory.hpp"
-#include "ode_solver/runge_kutta_ode_solver.h"
+//#include "ode_solver/runge_kutta_ode_solver.h"
+#include "ode_solver/explicit_ode_solver.h"
 #include "ode_solver/ode_solver_factory.h"
 #include <deal.II/base/table_handler.h>
 #include <string>
