+++ src/testing/time_refinement_study.cpp
@@ -3,7 +3,7 @@
 #include "flow_solver/flow_solver_cases/periodic_1D_unsteady.h"
 #include "physics/exact_solutions/exact_solution.h"
 #include "cmath"
-#include "ode_solver/runge_kutta_ode_solver.h"
+//#include "ode_solver/runge_kutta_ode_solver.h"
 
 namespace PHiLiP {
 namespace Tests {
